+++ tests/unit/Rules/ArrayValTest.php
@@ -11,69 +11,34 @@
 
 namespace Respect\Validation\Rules;
 
-class TestAccess extends \ArrayObject implements \ArrayAccess, \Countable, \Traversable
-{
-}
-
 /**
  * @group  rule
  * @covers Respect\Validation\Rules\ArrayVal
- * @covers Respect\Validation\Exceptions\ArrayValException
  */
-class ArrayValTest extends \PHPUnit_Framework_TestCase
+class ArrayValTest extends RuleTestCase
 {
-    protected $object;
-
-    protected function setUp()
+    public function providerForValidInput()
     {
-        $this->object = new ArrayVal();
-    }
+        $rule = new ArrayVal();
 
-    /**
-     * @dataProvider providerForArray
-     */
-    public function testValidArrayOrArrayObjectShouldReturnTrue($input)
-    {
-        $this->assertTrue($this->object->__invoke($input));
-        $this->assertTrue($this->object->assert($input));
-        $this->assertTrue($this->object->check($input));
-    }
-
-    /**
-     * @dataProvider providerForNotArray
-     * @expectedException Respect\Validation\Exceptions\ArrayValException
-     */
-    public function testNotArraysShouldThrowArrException($input)
-    {
-        $this->assertFalse($this->object->__invoke($input));
-        $this->assertFalse($this->object->assert($input));
-    }
-
-    public function providerForArray()
-    {
         return [
-            [[]],
-            [[1, 2, 3]],
-            [new TestAccess()],
+            [$rule, []],
+            [$rule, [1, 2, 3]],
+            [$rule, new \ArrayObject()],
         ];
-
-        $validator = v::alnum()->length(1, 10);
-
-        $validator = new \Respect\Validation\Rules\AllOf(
-            new Respect\Validation\Rules\Alnum(),
-            new Respect\Validation\Rules\Length(1, 10)
-        );
     }
 
-    public function providerForNotArray()
+    public function providerForInvalidInput()
     {
+        $rule = new ArrayVal();
+
         return [
-            [''],
-            [null],
-            [121],
-            [new \stdClass()],
-            [false],
-            ['aaa'],
+            [$rule, ''],
+            [$rule, null],
+            [$rule, 121],
+            [$rule, new \stdClass()],
+            [$rule, false],
+            [$rule, 'aaa'],
         ];
     }
 }